Login fragment update (#22)
* Login working
The atomic integer is now retrieved from the asynchronous operation in the read method in the Connect class

* Register working
Created a simple register layout and connected everything to the database, it checks for existing emails. Success = redirected to map (logged in). Currently no way to get back to the login screen

Co-authored-by: hansson0950 <emil.hansson@live.se>